services/s3-sink/requirements.txt
paho-mqtt==2.1.0
boto3==1.42.89
pandas==3.0.2
python-dotenv==1.2.2